refactor(form): rename InputText folder to Text
