+++ packages/css/src/agent.scss
@@ -1,4 +1,4 @@
 @use "./Button/Button.agent";
 @use "./Form/core/FormCore.agent.scss";
-@use "./Form/InputText/InputText.agent.scss";
+@use "./Form/Text/InputText.agent.scss";
 @use "./Form/InputTextarea/InputTextarea.agent.scss";
