+++ packages/react/src/agent.ts
@@ -2,8 +2,8 @@ import "@axa-fr/design-system-css/dist/common/icons.scss";
 import "@fontsource/source-sans-pro";
 
 export { ButtonAgent as Button } from "./Button/Button.agent";
-export { Text, TextInput } from "./Form/InputText";
 export { Select, SelectBase, SelectInput } from "./Form/Select";
+export { Text, TextInput } from "./Form/Text";
 export { Textarea, TextareaInput } from "./Form/Textarea";
 export {
   Field,
@@ -14,7 +14,8 @@ export {
   HelpMessage,
   InputList,
   MessageTypes,
-  getComponentClassName,
+  getComponentClassName
 } from "./Form/core";
 
 export { Number, NumberInput } from "./Form/Number";
+
